+++ packages/terra-draw/src/common.ts
@@ -74,7 +74,6 @@ export type SetCursor = (
 		| "crosshair"
 		| "pointer"
 		| "wait"
-		| "default"
 		| "move",
 ) => void;