+++ packages/terra-draw/src/test/test-adapter.ts
@@ -61,8 +61,7 @@ export class TerraDrawTestAdapter extends TerraDrawBaseAdapter {
 			| "grabbing"
 			| "crosshair"
 			| "pointer"
-			| "wait"
-			| "default",
+			| "wait",
 	): ReturnType<SetCursor> {
 		// pass
 	}
@@ -95,9 +94,8 @@ export class TerraDrawTestAdapter extends TerraDrawBaseAdapter {
  * custom draw modes for Terra Draw exclusively relying on the public API of the library.
  */
 
-interface TerraDrawTestModeOptions<
-	T extends CustomStyling,
-> extends TerraDrawExtend.BaseModeOptions<T> {
+interface TerraDrawTestModeOptions<T extends CustomStyling>
+	extends TerraDrawExtend.BaseModeOptions<T> {
 	customProperty: string;
 }